requirements.txt
alembic==1.4.3
aniso8601==8.0.0
apache-airflow==1.10.13
apache-airflow-backport-providers-google==2020.11.23
apispec==1.3.3
appdirs==1.4.4
argcomplete==1.12.2
argon2-cffi==20.1.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.1.5
cached-property==1.5.2
cachetools==4.1.1
cattrs==1.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.36
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
dnspython==2.0.0
docutils==0.16
email-validator==1.1.2
entrypoints==0.3
environment-kernels==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.0
Flask-Login==0.4.1
flask-marshmallow==0.14.0
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
gevent==20.9.0
gitdb==4.0.5
GitPython==3.1.11
google-ads==8.0.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-automl==1.0.1
google-cloud-bigquery==2.4.0
google-cloud-bigquery-datatransfer==1.1.1
google-cloud-bigquery-storage==2.1.0
google-cloud-bigtable==1.6.0
google-cloud-container==1.0.1
google-cloud-core==1.4.3
google-cloud-datacatalog==0.7.0
google-cloud-dataproc==1.1.1
google-cloud-dlp==1.0.0
google-cloud-kms==1.4.0
google-cloud-language==1.3.0
google-cloud-logging==1.15.1
google-cloud-memcache==0.2.0
google-cloud-monitoring==1.1.0
google-cloud-os-login==1.0.0
google-cloud-pubsub==1.7.0
google-cloud-redis==1.0.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.33.0
google-cloud-tasks==1.5.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphviz==0.15
greenlet==0.4.17
grpc-google-iam-v1==0.12.3
grpcio==1.33.2
grpcio-gcp==0.2.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==2.1.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.13
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json-merge-patch==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
kubernetes==11.0.0
lazy-object-proxy==1.5.1
libcst==0.3.14
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.1
mistune==0.8.4
mock==4.0.3
mypy-extensions==0.4.3
natsort==7.1.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
pandas-gbq==0.14.1
pandocfilters==1.4.2
parso==0.7.1
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prison==0.1.3
prometheus-client==0.8.0
prompt-toolkit==3.0.6
proto-plus==1.11.0
protobuf==3.14.0
psutil==5.7.3
ptyprocess==0.6.0
py==1.9.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.1.0
Pygments==2.7.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytest==6.1.2
pytest-mock==3.4.0
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
setproctitle==1.2.1
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
sqlite-bro==0.9.1
tabulate==0.8.7
tenacity==4.12.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
thrift==0.13.0
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.0.31
wcwidth==0.2.5
webargs==6.1.1
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
WTForms==2.3.3
xlrd==1.2.0
zipp==3.4.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.interface==5.2.0
